v0.82 - Close to Happy
I'm very close to being happy with a full version. For now the portfolio remains my biggest issue as I like the animations put in place but they're still not good with small screens.

Also, there's a lot of unused space in portfolio when a language is not selected. I'd like to fill that with something.

Once those two things are changed, I'll be comfortable to move on to v0.9. A full V1.0 will only be once all information is provided and my projects have been uploaded.